Install yarn. Needed before you can start RoR.
Hi,
I followed the tutorial on [freecodecamp](https://www.freecodecamp.org/news/painless-rails-development-environment-setup-with-docker/#building-and-running-the-container) but there is one step missing involving `yarn`.

You will receive an error when starting the server because webpack is missing and you need yarn to install it. yarn is not included in this dockerfile.

This pr fix the problem.